+++ crates/fricon-ui/frontend/src/features/charts/hooks/useWebGLChart.test.tsx
@@ -194,7 +194,7 @@ function xySeries(id: string, label: string, points: [number, number][]) {
     id,
     label,
     pointCount: points.length,
-    values: Float32Array.from(points.flat()),
+    values: Float64Array.from(points.flat()),
   };
 }
 
@@ -207,7 +207,7 @@ function xyzSeries(
     id,
     label,
     pointCount: points.length,
-    values: Float32Array.from(points.flat()),
+    values: Float64Array.from(points.flat()),
   };
 }
