+++ crates/fricon-ui/frontend/src/features/charts/ui/ChartWrapper.test.tsx
@@ -203,7 +203,7 @@ function xySeries(id: string, label: string, points: [number, number][]) {
     id,
     label,
     pointCount: points.length,
-    values: Float32Array.from(points.flat()),
+    values: Float64Array.from(points.flat()),
   };
 }
 
@@ -216,6 +216,6 @@ function xyzSeries(
     id,
     label,
     pointCount: points.length,
-    values: Float32Array.from(points.flat()),
+    values: Float64Array.from(points.flat()),
   };
 }
